requirements.txt
pymongo
gradio
google-genai
python-dotenv
pytz